Unify .NET/.NET Framework tests
