+++ test/OpenTelemetry.Instrumentation.SqlClient.Tests/SqlClientTestCases.json
@@ -20,6 +20,13 @@
       "server.port": null,
       "spanName": "select sys.databases"
     },
+    "expectedNetFramework": {
+      "db.operation.name": null,
+      "db.query.summary": "select sys.databases",
+      "db.query.text": "select * from sys.databases",
+      "db.stored_procedure.name": null,
+      "spanName": "select sys.databases"
+    },
     "expectedOldConventions": {
       "db.mssql.instance_name": "MSSQLLocalDB",
       "db.name": "master",
@@ -48,6 +55,13 @@
       "server.port": null,
       "spanName": "EXECUTE SP_GetOrders"
     },
+    "expectedNetFramework": {
+      "db.operation.name": null,
+      "db.query.summary": null,
+      "db.query.text": "SP_GetOrders",
+      "db.stored_procedure.name": null,
+      "spanName": "MSSQLLocalDB.master"
+    },
     "expectedOldConventions": {
       "db.mssql.instance_name": "MSSQLLocalDB",
       "db.name": "master",
@@ -74,6 +88,13 @@
       "error.type": null,
       "server.address": "(localdb)",
       "server.port": 1434,
+      "spanName": "select sys.databases"
+    },
+    "expectedNetFramework": {
+      "db.operation.name": null,
+      "db.query.summary": "select sys.databases",
+      "db.query.text": "select * from sys.databases",
+      "db.stored_procedure.name": null,
       "spanName": "select sys.databases"
     },
     "expectedOldConventions": {
